moved voice keyer code to a separate file
